Tradução de nomes de arquivos e padronizações
