+++ criando-uma-aplicacao/servidor-http/v1/main.go
@@ -6,7 +6,7 @@ import (
 )
 
 func main() {
-	tratador := http.HandlerFunc(JogadorServidor)
+	tratador := http.HandlerFunc(ServidorJogador)
 	if err := http.ListenAndServe(":5000", tratador); err != nil {
 		log.Fatalf("não foi possível ouvir na porta 5000 %v", err)
 	}
