--- criando-uma-aplicacao/servidor-http/v1/server.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"net/http"
-)
-
-// JogadorServidor retorna valor fixo "20" para qualquer chamada
-func JogadorServidor(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprint(w, "20")
-}
