+++ criando-uma-aplicacao/servidor-http/v2/main.go
@@ -5,16 +5,16 @@ import (
 	"net/http"
 )
 
-// JogadorArmazenamentoEmMemoria armazena na memória os dados sobre os jogadores
-type JogadorArmazenamentoEmMemoria struct{}
+// ArmazenamentoJogadorEmMemoria armazena na memória os dados sobre os jogadores
+type ArmazenamentoJogadorEmMemoria struct{}
 
 // ObterPontuacaoJogador obtém as pontuações para um jogador
-func (i *JogadorArmazenamentoEmMemoria) ObterPontuacaoJogador(nome string) int {
+func (i *ArmazenamentoJogadorEmMemoria) ObterPontuacaoJogador(nome string) int {
 	return 123
 }
 
 func main() {
-	servidor := &JogadorServidor{&JogadorArmazenamentoEmMemoria{}}
+	servidor := &ServidorJogador{&ArmazenamentoJogadorEmMemoria{}}
 
 	if err := http.ListenAndServe(":5000", servidor); err != nil {
 		log.Fatalf("não foi possível ouvir na porta 5000 %v", err)
