+++ criando-uma-aplicacao/servidor-http/v2/servidor_test.go
@@ -7,32 +7,26 @@ import (
 	"testing"
 )
 
-type EsbocoJogadorArmazenamento struct {
-	pontuacoes        map[string]int
-	registrosVitorias []string
+type EsbocoArmazenamentoJogador struct {
+	pontuacoes map[string]int
 }
 
-func (s *EsbocoJogadorArmazenamento) ObterPontuacaoJogador(nome string) int {
-	pontuacao := s.pontuacoes[nome]
+func (e *EsbocoArmazenamentoJogador) ObterPontuacaoJogador(nome string) int {
+	pontuacao := e.pontuacoes[nome]
 	return pontuacao
 }
 
-func (s *EsbocoJogadorArmazenamento) RegistrarVitoria(nome string) {
-	s.registrosVitorias = append(s.registrosVitorias, nome)
-}
-
 func TestObterJogadores(t *testing.T) {
-	armazenamento := EsbocoJogadorArmazenamento{
+	armazenamento := EsbocoArmazenamentoJogador{
 		map[string]int{
 			"Maria": 20,
 			"Pedro": 10,
 		},
-		nil,
 	}
-	servidor := &JogadorServidor{&armazenamento}
+	servidor := &ServidorJogador{&armazenamento}
 
 	t.Run("obter pontuação de Maria", func(t *testing.T) {
-		requisicao := novaRequisicaoPontuacaoGet("Pepper")
+		requisicao := novaRequisicaoPontuacaoGet("Maria")
 		resposta := httptest.NewRecorder()
 
 		servidor.ServeHTTP(resposta, requisicao)
@@ -42,7 +36,7 @@ func TestObterJogadores(t *testing.T) {
 	})
 
 	t.Run("obter pontuação de Pedro", func(t *testing.T) {
-		requisicao := novaRequisicaoPontuacaoGet("Floyd")
+		requisicao := novaRequisicaoPontuacaoGet("Pedro")
 		resposta := httptest.NewRecorder()
 
 		servidor.ServeHTTP(resposta, requisicao)
@@ -58,33 +52,6 @@ func TestObterJogadores(t *testing.T) {
 		servidor.ServeHTTP(resposta, requisicao)
 
 		verificarRespostaCodigoStatus(t, resposta.Code, http.StatusNotFound)
-	})
-}
-
-func TestArmazenamentoVitorias(t *testing.T) {
-	armazenamento := EsbocoJogadorArmazenamento{
-		map[string]int{},
-		nil,
-	}
-	servidor := &JogadorServidor{&armazenamento}
-
-	t.Run("registra vitorias na chamada ao método HTTP POST", func(t *testing.T) {
-		jogador := "Pepper"
-
-		requisicao, _ := http.NewRequest(http.MethodPost, fmt.Sprintf("/jogadores/%s", jogador), nil)
-		resposta := httptest.NewRecorder()
-
-		servidor.ServeHTTP(resposta, requisicao)
-
-		verificarRespostaCodigoStatus(t, resposta.Code, http.StatusAccepted)
-
-		if len(armazenamento.registrosVitorias) != 1 {
-			t.Fatalf("verifiquei %d chamadas a RegistrarVitoria, esperava %d", len(armazenamento.registrosVitorias), 1)
-		}
-
-		if armazenamento.registrosVitorias[0] != jogador {
-			t.Errorf("não registrou o vencedor corretamente, recebi '%s', esperava '%s'", armazenamento.registrosVitorias[0], jogador)
-		}
 	})
 }
