+++ criando-uma-aplicacao/servidor-http/v4/servidor.go
@@ -1,0 +1,43 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+)
+
+// ArmazenamentoJogador armazena as pontuacoes dos jogadores
+type ArmazenamentoJogador interface {
+	ObterPontuacaoJogador(nome string) int
+	RegistrarVitoria(nome string)
+}
+
+// ServidorJogador é uma interface HTTP para os dados dos jogadores
+type ServidorJogador struct {
+	armazenamento ArmazenamentoJogador
+}
+
+func (s *ServidorJogador) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	jogador := r.URL.Path[len("/jogadores/"):]
+
+	switch r.Method {
+	case http.MethodPost:
+		s.registrarVitoria(w, jogador)
+	case http.MethodGet:
+		s.mostrarPontuacao(w, jogador)
+	}
+}
+
+func (s *ServidorJogador) mostrarPontuacao(w http.ResponseWriter, jogador string) {
+	pontuacao := s.armazenamento.ObterPontuacaoJogador(jogador)
+
+	if pontuacao == 0 {
+		w.WriteHeader(http.StatusNotFound)
+	}
+
+	fmt.Fprint(w, pontuacao)
+}
+
+func (s *ServidorJogador) registrarVitoria(w http.ResponseWriter, jogador string) {
+	s.armazenamento.RegistrarVitoria(jogador)
+	w.WriteHeader(http.StatusAccepted)
+}
