+++ criando-uma-aplicacao/servidor-http/v5/servidor_integration_test.go
@@ -7,8 +7,8 @@ import (
 )
 
 func TestRegistrarVitoriasEBuscarEstasVitorias(t *testing.T) {
-	armazenamento := CriarJogadorArmazenamentoNaMemoria()
-	servidor := JogadorServidor{armazenamento}
+	armazenamento := NovoArmazenamentoJogadorEmMemoria()
+	servidor := ServidorJogador{armazenamento}
 	jogador := "Maria"
 
 	servidor.ServeHTTP(httptest.NewRecorder(), novaRequisicaoPontuacaoPost(jogador))
